[hotfix][table] Move generate runner functions to MatchCodeGenerator companion object
